+++ __tests__/__integration__/__scripts__/imperative_zosmf_create_profile.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -e
-
-echo "================ZOSMF CREATE PROFILE==============="
-zowe profiles create zosmf $*
-if [ $? -gt 0 ]
-then
-    exit $?
-fi