+++ packages/imperative/__tests__/__integration__/imperative/__tests__/__integration__/cli/config/convert-profiles/__scripts__/delete_profiles.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+# Delete all available profiles, but leave the profile type definitions
+for profile in profiles/base/test.yaml profiles/secured/test.yaml profiles/v1profile/myv1profile.yaml; do
+    if [ -e $profile ]; then
+        rm $profile
+    fi
+done