+++ packages/imperative/__tests__/src/example_clis/with_profiles/__integration__/__scripts__/copy_auto_gen_profiles.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+# include zowe-cli\__tests__\__scripts__\exitOnFailure function
+myScriptDir=`dirname $0`
+. $myScriptDir/../../../../../../../../__tests__/__scripts__/exitOnFailure.sh
+
+# copy pre-existing profiles to test directory
+cp -r $myScriptDir/../__resources__/autoGenProfiles profiles
+exitOnFailure "Failed to copy test profile." $?