+++ packages/imperative/__tests__/src/packages/profiles/__integration__/BasicProfileManager.integration.test.ts
@@ -10,12 +10,16 @@
 */
 
 import * as TestUtil from "../../../TestUtil";
+import { ITestEnvironment } from "../../../../__src__/environment/doc/response/ITestEnvironment";
+import { SetupTestEnvironment } from "../../../../__src__/environment/SetupTestEnvironment";
 import { inspect } from "util";
 import { TestLogger } from "../../../../src/TestLogger";
 import { ProfileUtils } from "../../../../../src/profiles";
-import { BANANA_AGE, getConfig, PROFILE_TYPE } from "../src/constants/BasicProfileManagerTestConstants";
+import { getConfig, PROFILE_TYPE } from "../src/constants/BasicProfileManagerTestConstants";
 
-describe("Imperative should allow CLI implementations to configure their own profiles and types", function () {
+let TEST_ENVIRONMENT: ITestEnvironment;
+
+describe("Imperative should allow CLI implementations to read their own profiles and types", function () {
     const mainModule = process.mainModule;
     const loadChangingDependencies = () => {
         return {
@@ -27,31 +31,42 @@ describe("Imperative should allow CLI implementations to configure their own pro
 
     let {Imperative, ImperativeError, ImperativeConfig} = loadChangingDependencies();
 
+    beforeAll(async () => {
+        (process.mainModule as any) = {
+            filename: __filename
+        };
+
+        TEST_ENVIRONMENT = await SetupTestEnvironment.createTestEnv({
+            cliHomeEnvVar: "CMD_CLI_CLI_HOME",
+            testName: "basic_profile_mgr"
+        });
+
+        // copy existing profiles into test directory
+        const response = TestUtil.runCliScript(__dirname + "/__scripts__/copy_profiles.sh", TEST_ENVIRONMENT.workingDir);
+        expect(response.stderr.toString()).toBe("");
+        expect(response.status).toBe(0);
+    });
+
     // Initialize imperative before each test
     beforeEach(() => {
         jest.resetModules();
         ({Imperative, ImperativeError, ImperativeConfig} = loadChangingDependencies());
     });
 
-    beforeAll(() => {
-        (process.mainModule as any) = {
-            filename: __filename
-        };
-    });
-
     afterAll(() => {
         process.mainModule = mainModule;
+        TestUtil.rimraf(TEST_ENVIRONMENT.workingDir);
     });
 
-    it("should be able to create a profile type and retrieve all defined types after init", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
+    it("should be able to retrieve all defined types after init", async function () {
+        const config = getConfig(TEST_ENVIRONMENT.workingDir);
         await Imperative.init(config);
         expect(ProfileUtils.getAllTypeNames(ImperativeConfig.instance.loadedConfig.profiles).length).toEqual(Object.keys(PROFILE_TYPE).length);
         expect(ProfileUtils.getAllTypeNames(ImperativeConfig.instance.loadedConfig.profiles)).toContain("banana");
     });
 
-    it("should be receive a failure message when attempting to load a profile that doesn't exist", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
+    it("should receive a failure message when attempting to load a profile that doesn't exist", async function () {
+        const config = getConfig(TEST_ENVIRONMENT.workingDir);
         await Imperative.init(config);
         let error;
         try {
@@ -64,72 +79,30 @@ describe("Imperative should allow CLI implementations to configure their own pro
         expect(error).toBeInstanceOf(ImperativeError);
     });
 
-    it("should receive a failure message when attempting to create an ill-formed profile", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
-        await Imperative.init(config);
-        let error;
-        try {
-            const response = await Imperative.api.profileManager("banana").save({
-                name: "illformed",
-                profile: {
-                    fail: "this"
-                }
-            });
-            TestLogger.info(response.message);
-        } catch (e) {
-            error = e;
-            TestLogger.info(error.message);
-        }
-        expect(error).toBeInstanceOf(ImperativeError);
-    });
+    it("should be able to load a specific profile", async function () {
+        const config = getConfig(TEST_ENVIRONMENT.workingDir);
 
-    it("should be able to create a basic profile", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
         await Imperative.init(config);
-        const response = await Imperative.api.profileManager("banana").save({
-            name: "legit",
-            profile: {
-                age: BANANA_AGE
-            }
-        });
-        TestLogger.info(`Profile create success response: ${inspect(response, {depth: null})}`);
-        expect(response.message).toContain(`Profile ("legit" of type "banana") successfully written:`);
-        expect(response.overwritten).toEqual(false);
-        expect(response.path).toContain("legit.yaml");
-    });
-
-    it("should be able to create a basic profile and load that profile", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
-        await Imperative.init(config);
-        const saveResponse = await Imperative.api.profileManager("banana").save({
-            name: "legit",
-            profile: {age: BANANA_AGE},
-            overwrite: true,
-            updateDefault: true
-        });
-
-        TestLogger.info("Save response: " + saveResponse.message);
         const loadResponse = await Imperative.api.profileManager("banana").load({name: "legit"});
         TestLogger.info(`Profile loaded success response: ${inspect(loadResponse, {depth: null})}`);
-        expect(loadResponse.name).toEqual("legit");
+
+        expect(loadResponse.message).toEqual('Profile "legit" of type "banana" loaded successfully.');
         expect(loadResponse.type).toEqual("banana");
+        expect(loadResponse.name).toEqual("legit");
         expect(loadResponse.profile).toBeDefined();
-        expect(loadResponse.profile).toMatchSnapshot();
+        expect(loadResponse.profile).toEqual({"age": 1000});
     });
 
-    it("should be able to create a basic profile and load as the default", async function () {
-        const config = getConfig(TestUtil.createUniqueTestDataDir("profile-manager-initialize"));
+    it("should be able to load a default profile", async function () {
+        const config = getConfig(TEST_ENVIRONMENT.workingDir);
+
         await Imperative.init(config);
-        const response = await Imperative.api.profileManager("banana").save({
-            name: "legit",
-            profile: {age: BANANA_AGE},
-            overwrite: true,
-            updateDefault: true
-        });
-        TestLogger.info(`Profile create success response: ${inspect(response, {depth: null})}`);
-        const load = await Imperative.api.profileManager("banana").load({loadDefault: true});
-        expect(load.type).toEqual("banana");
-        expect(load.profile).toBeDefined();
-        expect(load.profile).toMatchSnapshot();
+        const loadResponse = await Imperative.api.profileManager("banana").load({loadDefault: true});
+
+        expect(loadResponse.message).toEqual('Profile "legit" of type "banana" loaded successfully.');
+        expect(loadResponse.type).toEqual("banana");
+        expect(loadResponse.name).toEqual("legit");
+        expect(loadResponse.profile).toBeDefined();
+        expect(loadResponse.profile).toEqual({ "age": 1000 });
     });
 });