+++ packages/imperative/src/profiles/index.ts
@@ -23,7 +23,6 @@ export * from "./src/doc/api/IProfileManagerFactory";
 export * from "./src/doc/parms/IDeleteProfile";
 export * from "./src/doc/parms/ILoadProfile";
 export * from "./src/doc/parms/IProfileManager";
-export * from "./src/doc/parms/IProfileManagerInit";
 export * from "./src/doc/parms/ISaveProfile";
 export * from "./src/doc/parms/ISaveProfileFromCliArgs";
 export * from "./src/doc/parms/ISetDefaultProfile";
@@ -36,7 +35,6 @@ export * from "./src/doc/parms/IValidateProfileWithSchema";
 export * from "./src/doc/api/IProfileManagerFactory";
 
 export * from "./src/doc/response/IProfileDeleted";
-export * from "./src/doc/response/IProfileInitialized";
 export * from "./src/doc/response/IProfileLoaded";
 export * from "./src/doc/response/IProfileSaved";
 export * from "./src/doc/response/IProfileUpdated";