+++ packages/zosmf/__tests__/__system__/methods/ListDefinedSystems.system.test.ts
@@ -92,7 +92,10 @@ describe("List Defined Systems Api", () => {
 
             expect(error).toBeTruthy();
             expect(response).toBeFalsy();
-            expect(error.message).toMatch(/(Error: getaddrinfo).*(badHost)/);
+            const jsonCauseErrors = error.causeErrors;
+            expect(jsonCauseErrors.code).toEqual("ENOTFOUND");
+            expect(jsonCauseErrors.syscall).toEqual("getaddrinfo");
+            expect(jsonCauseErrors.hostname).toEqual(badHostName);
         });
 
         it("should return with proper message for invalid port", async () => {
@@ -118,7 +121,9 @@ describe("List Defined Systems Api", () => {
 
             expect(error).toBeTruthy();
             expect(response).toBeFalsy();
-            expect(error.message).toMatch(/Error: (connect|read) (ECONNREFUSED|ECONNRESET)/);
+            const jsonCauseErrors = error.causeErrors;
+            expect(jsonCauseErrors.code).toMatch(/(ECONNREFUSED|ECONNRESET)/);
+            expect(jsonCauseErrors.syscall).toMatch(/(connect|read)/);
         });
     });
 });